# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Maryland Representatives Civic App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RepresentativeLevel =
  (typeof RepresentativeLevel)[keyof typeof RepresentativeLevel];

export const RepresentativeLevel = {
  federal: "federal",
  state: "state",
  local: "local",
} as const;

export interface Representative {
  name: string;
  office: string;
  party?: string;
  phone?: string;
  photoUrl?: string;
  level: RepresentativeLevel;
  chamber?: string;
  urls?: string[];
  emails?: string[];
  bioguideId?: string;
  openstatesId?: string;
  division?: string;
}

export interface RepresentativesResponse {
  address: string;
  representatives: Representative[];
  normalizedAddress?: string;
}

export interface Bill {
  id: string;
  title: string;
  number?: string;
  congress?: string;
  introducedDate?: string;
  latestAction?: string;
  latestActionDate?: string;
  sponsors?: string[];
  url?: string;
  status?: string;
  chamber?: string;
}

export interface SponsorInfo {
  name: string;
  party?: string;
  state?: string;
  bioguideId?: string;
}

export interface CommitteeRef {
  name: string;
  chamber?: string;
  committeeCode?: string;
}

export interface BillAction {
  date: string;
  text: string;
  type?: string;
}

export interface BillDetail {
  id: string;
  title: string;
  number?: string;
  congress?: string;
  introducedDate?: string;
  summary?: string;
  status?: string;
  latestAction?: string;
  latestActionDate?: string;
  sponsors?: SponsorInfo[];
  cosponsors?: SponsorInfo[];
  committees?: CommitteeRef[];
  actions?: BillAction[];
  url?: string;
}

export interface BillsListResponse {
  bills: Bill[];
  totalCount?: number;
  offset?: number;
}

export interface VoteRecord {
  billId?: string;
  billTitle: string;
  billNumber?: string;
  date: string;
  position: string;
  question?: string;
  result?: string;
  description?: string;
}

export interface VotesListResponse {
  votes: VoteRecord[];
  totalCount?: number;
  offset?: number;
}

export interface CommitteeMember {
  name: string;
  party?: string;
  state?: string;
  rank?: string;
  bioguideId?: string;
}

export interface Committee {
  name: string;
  chamber?: string;
  committeeCode?: string;
  rank?: string;
  title?: string;
  members?: CommitteeMember[];
}

export interface CommitteesListResponse {
  committees: Committee[];
}

export interface FederalMemberDetail {
  bioguideId: string;
  name: string;
  party?: string;
  state?: string;
  chamber?: string;
  district?: string;
  phone?: string;
  website?: string;
  photoUrl?: string;
  terms?: number;
  inOffice?: boolean;
  nextElection?: string;
}

export interface StateMemberDetail {
  id: string;
  name: string;
  party?: string;
  chamber?: string;
  district?: string;
  email?: string;
  phone?: string;
  photoUrl?: string;
  openstatesUrl?: string;
}

export interface StateBill {
  id: string;
  identifier?: string;
  title: string;
  session?: string;
  chamber?: string;
  status?: string;
  introducedDate?: string;
  latestAction?: string;
  sponsors?: string[];
  url?: string;
}

export interface StateBillVote {
  date: string;
  chamber?: string;
  result: string;
  yesCount?: number;
  noCount?: number;
  absentCount?: number;
}

export interface StateBillDetail {
  id: string;
  identifier?: string;
  title: string;
  session?: string;
  chamber?: string;
  status?: string;
  introducedDate?: string;
  summary?: string;
  sponsors?: SponsorInfo[];
  cosponsors?: SponsorInfo[];
  committees?: CommitteeRef[];
  actions?: BillAction[];
  votes?: StateBillVote[];
  url?: string;
}

export interface StateBillsListResponse {
  bills: StateBill[];
  totalCount?: number;
  offset?: number;
}

export interface StateVoteRecord {
  billId?: string;
  billTitle: string;
  billIdentifier?: string;
  date: string;
  position: string;
  result?: string;
}

export interface StateVotesListResponse {
  votes: StateVoteRecord[];
}

export type DonorEntryType =
  (typeof DonorEntryType)[keyof typeof DonorEntryType];

export const DonorEntryType = {
  pac: "pac",
  individual: "individual",
  other: "other",
} as const;

export interface DonorEntry {
  name: string;
  total: number;
  type?: DonorEntryType;
  industry?: string;
}

export interface FinanceResponse {
  candidateId: string;
  candidateName: string;
  cycle?: number;
  totalRaised?: number;
  totalSpent?: number;
  cashOnHand?: number;
  topDonors: DonorEntry[];
  topIndustries?: DonorEntry[];
}

export interface FinanceCandidate {
  id: string;
  name: string;
  party?: string;
  state?: string;
  office?: string;
  cycle?: number;
}

export interface FinanceSearchResponse {
  candidates: FinanceCandidate[];
}

export type GetRepresentativesByAddressParams = {
  address: string;
};

export type GetFederalMemberBillsParams = {
  type?: GetFederalMemberBillsType;
  offset?: number;
  limit?: number;
};

export type GetFederalMemberBillsType =
  (typeof GetFederalMemberBillsType)[keyof typeof GetFederalMemberBillsType];

export const GetFederalMemberBillsType = {
  sponsored: "sponsored",
  cosponsored: "cosponsored",
} as const;

export type GetFederalMemberVotesParams = {
  offset?: number;
  limit?: number;
};

export type GetFederalBillsParams = {
  chamber?: GetFederalBillsChamber;
  offset?: number;
  limit?: number;
};

export type GetFederalBillsChamber =
  (typeof GetFederalBillsChamber)[keyof typeof GetFederalBillsChamber];

export const GetFederalBillsChamber = {
  house: "house",
  senate: "senate",
  both: "both",
} as const;

export type GetStateMemberBillsParams = {
  type?: GetStateMemberBillsType;
};

export type GetStateMemberBillsType =
  (typeof GetStateMemberBillsType)[keyof typeof GetStateMemberBillsType];

export const GetStateMemberBillsType = {
  sponsored: "sponsored",
  cosponsored: "cosponsored",
} as const;

export type GetStateBillsParams = {
  chamber?: GetStateBillsChamber;
  offset?: number;
  limit?: number;
};

export type GetStateBillsChamber =
  (typeof GetStateBillsChamber)[keyof typeof GetStateBillsChamber];

export const GetStateBillsChamber = {
  upper: "upper",
  lower: "lower",
} as const;

export type GetCandidateFinanceParams = {
  cycle?: number;
};

export type SearchCandidateFinanceParams = {
  name: string;
  state?: string;
};