# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maryland Representatives Civic App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /federal/members/{bioguideId}:
    get:
      operationId: getFederalMember
      tags: [federal]
      summary: Get a federal member's details
      parameters:
        - name: bioguideId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FederalMemberDetail"
  /federal/members/{bioguideId}/bills:
    get:
      operationId: getFederalMemberBills
      tags: [federal]
      summary: Get bills sponsored or cosponsored by a federal member
      parameters:
        - name: bioguideId
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [sponsored, cosponsored]
            default: sponsored
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Bills list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillsListResponse"
  /federal/members/{bioguideId}/votes:
    get:
      operationId: getFederalMemberVotes
      tags: [federal]
      summary: Get voting record for a federal member
      parameters:
        - name: bioguideId
          in: path
          required: true
          schema:
            type: string
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Votes list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VotesListResponse"
  /federal/members/{bioguideId}/committees:
    get:
      operationId: getFederalMemberCommittees
      tags: [federal]
      summary: Get committee memberships for a federal member
      parameters:
        - name: bioguideId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Committees list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitteesListResponse"
  /federal/bills:
    get:
      operationId: getFederalBills
      tags: [federal]
      summary: Get bills currently being considered in Congress
      parameters:
        - name: chamber
          in: query
          schema:
            type: string
            enum: [house, senate, both]
            default: both
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Bills list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillsListResponse"
  /federal/bills/{congress}/{billType}/{billNumber}:
    get:
      operationId: getFederalBillDetail
      tags: [federal]
      summary: Get detailed info about a specific federal bill
      parameters:
        - name: congress
          in: path
          required: true
          schema:
            type: string
        - name: billType
          in: path
          required: true
          schema:
            type: string
        - name: billNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bill detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDetail"
...
  /finance/{candidateId}:
    get:
      operationId: getCandidateFinance
      tags: [finance]
      summary: Get PAC and donor information for a candidate
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: string
        - name: cycle
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Finance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceResponse"
  /finance/search:
    get:
      operationId: searchCandidateFinance
      tags: [finance]
      summary: Search for a candidate's FEC finance data by name
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Finance search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinanceSearchResponse"
components:
  schemas:
...
    Bill:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        number:
          type: string
        congress:
          type: string
        introducedDate:
          type: string
        latestAction:
          type: string
        latestActionDate:
          type: string
        sponsors:
          type: array
          items:
            type: string
        url:
          type: string
        status:
          type: string
        chamber:
          type: string
      required: [id, title]
    BillDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        number:
          type: string
        congress:
          type: string
        introducedDate:
          type: string
        summary:
          type: string
        status:
          type: string
        latestAction:
          type: string
        latestActionDate:
          type: string
        sponsors:
          type: array
          items:
            $ref: "#/components/schemas/SponsorInfo"
        cosponsors:
          type: array
          items:
            $ref: "#/components/schemas/SponsorInfo"
        committees:
          type: array
          items:
            $ref: "#/components/schemas/CommitteeRef"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/BillAction"
        url:
          type: string
      required: [id, title]
    SponsorInfo:
      type: object
      properties:
        name:
          type: string
        party:
          type: string
        state:
          type: string
        bioguideId:
          type: string
      required: [name]
    CommitteeRef:
      type: object
      properties:
        name:
          type: string
        chamber:
          type: string
        committeeCode:
          type: string
      required: [name]
    BillAction:
      type: object
      properties:
        date:
          type: string
        text:
          type: string
        type:
          type: string
      required: [date, text]
    BillsListResponse:
      type: object
      properties:
        bills:
          type: array
          items:
            $ref: "#/components/schemas/Bill"
        totalCount:
          type: integer
        offset:
          type: integer
      required: [bills]
    VoteRecord:
      type: object
      properties:
        billId:
          type: string
        billTitle:
          type: string
        billNumber:
          type: string
        date:
          type: string
        position:
          type: string
        question:
          type: string
        result:
          type: string
        description:
          type: string
      required: [billTitle, date, position]
    VotesListResponse:
      type: object
      properties:
        votes:
          type: array
          items:
            $ref: "#/components/schemas/VoteRecord"
        totalCount:
          type: integer
        offset:
          type: integer
      required: [votes]
    Committee:
      type: object
      properties:
        name:
          type: string
        chamber:
          type: string
        committeeCode:
          type: string
        rank:
          type: string
        title:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/CommitteeMember"
      required: [name]
    CommitteeMember:
      type: object
      properties:
        name:
          type: string
        party:
          type: string
        state:
          type: string
        rank:
          type: string
        bioguideId:
          type: string
      required: [name]
    CommitteesListResponse:
      type: object
      properties:
        committees:
          type: array
          items:
            $ref: "#/components/schemas/Committee"
      required: [committees]
    FederalMemberDetail:
      type: object
      properties:
        bioguideId:
          type: string
        name:
          type: string
        party:
          type: string
        state:
          type: string
        chamber:
          type: string
        district:
          type: string
        phone:
          type: string
        website:
          type: string
        photoUrl:
          type: string
        terms:
          type: integer
        inOffice:
          type: boolean
        nextElection:
          type: string
      required: [bioguideId, name]
...
    DonorEntry:
      type: object
      properties:
        name:
          type: string
        total:
          type: number
        type:
          type: string
          enum: [pac, individual, other]
        industry:
          type: string
      required: [name, total]
    FinanceResponse:
      type: object
      properties:
        candidateId:
          type: string
        candidateName:
          type: string
        cycle:
          type: integer
        totalRaised:
          type: number
        totalSpent:
          type: number
        cashOnHand:
          type: number
        topDonors:
          type: array
          items:
            $ref: "#/components/schemas/DonorEntry"
        topIndustries:
          type: array
          items:
            $ref: "#/components/schemas/DonorEntry"
      required: [candidateId, candidateName, topDonors]
    FinanceCandidate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        party:
          type: string
        state:
          type: string
        office:
          type: string
        cycle:
          type: integer
      required: [id, name]
    FinanceSearchResponse:
      type: object
      properties:
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/FinanceCandidate"
      required: [candidates]